backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
python-multipart
requests-toolbelt
boto3
